+++ gen/org/durka/hallmonitor/BuildConfig.java
@@ -1,0 +1,6 @@
+/** Automatically generated file. DO NOT MODIFY */
+package org.durka.hallmonitor;
+
+public final class BuildConfig {
+    public final static boolean DEBUG = false;
+}